Fix cyclical dependency between redux/fs/Load.tsx and redux/fs/index.ts
There is a cyclical dependency between redux/fs/Load.tsx and
redux/fs/index.ts, where one file imports the other. This is because the
`LoadRaw` function in Load.tsx is using the `loadFile` function in
index.ts. The latter function doesn't appear to be used anywhere else
or exported, so just move it into Load.tsx to break the cyclical
dependency.

This was causing the webapp to not load and instead fail with the
following error:

```
Uncaught ReferenceError: Cannot access '__WEBPACK_DEFAULT_EXPORT__' before initialization
```

Fixes #2031
