+++ src/redux/fs/Load.tsx
@@ -1,7 +1,6 @@
 import { Button, ButtonProps } from '@material-ui/core'
 import { ChangeEvent, PropsWithChildren, ReactElement } from 'react'
 import { intentionallyFloat } from 'ustaxes/core/util'
-import { loadFile } from '.'
 
 interface LoadProps<S> {
   handleData: (s: S) => void
@@ -9,6 +8,27 @@ interface LoadProps<S> {
 
 interface Accept {
   accept?: string
+}
+
+export const loadFile = async (file: File): Promise<string> => {
+  const fileReader: FileReader = new FileReader()
+
+  return new Promise((resolve, reject) => {
+    fileReader.onload = function (
+      this: FileReader,
+      e: ProgressEvent<FileReader>
+    ): void {
+      e.preventDefault()
+      // type known here because of readAsText
+      const contents: string | null = fileReader.result as string | null
+      if (contents === null) {
+        reject('file contents were null')
+      } else {
+        resolve(contents)
+      }
+    }
+    fileReader.readAsText(file)
+  })
 }
 
 export const LoadRaw = (
